+++ packages/components/src/Cards/ButtonCard/index.tsx
@@ -31,7 +31,12 @@ export const StyledButtonCard = styled.button`
   cursor: pointer;
   background-color: var(--ha-primary-background);
   box-shadow: 0px 2px 4px rgba(0, 0, 0, 0.1);
-  transition: all 0.2s cubic-bezier(0.06, 0.67, 0.37, 0.99);
+  transition: var(--ha-transition-duration) var(--ha-easing);
+  transition-property: box-shadow, transform;
+
+  &:active {
+    transform: translateZ(10px) scale(0.98);
+  }
 
   &:hover,
   &:focus,
@@ -159,7 +164,7 @@ export interface ButtonCardProps<
   onClick?: (entity: HassEntityWithApi<ExtractDomain<E>>) => void;
   /** Optional active param, By default this is updated via home assistant */
   active?: boolean;
-  /** the layout of the button card, this changes slightly, just preferences really */
+  /** the layout of the button card, this changes slightly, just preferences really @default default */
   layout?: "default" | "slim";
 }