+++ packages/components/src/Cards/TimeCard/TimeCard.stories.tsx
@@ -1,0 +1,47 @@
+import type { Meta, StoryObj } from "@storybook/react";
+import { ThemeProvider, TimeCard } from "@components";
+import type { TimeCardProps } from "@components";
+import { HassConnect } from "@stories/HassConnectFake";
+
+function Template(args?: Partial<TimeCardProps>) {
+  return (
+    <HassConnect hassUrl="http://localhost:8123">
+      <ThemeProvider />
+      <TimeCard {...args} />
+    </HassConnect>
+  );
+}
+
+function WithoutDate(args?: Partial<TimeCardProps>) {
+  return (
+    <div>
+      <h2>TimeCard without the date</h2>
+      <Template includeDate={false} {...args} />
+    </div>
+  );
+}
+
+export default {
+  title: "COMPONENTS/Cards/TimeCard",
+  component: TimeCard,
+  tags: ["autodocs"],
+  parameters: {
+    width: "100%",
+  },
+  argTypes: {
+    title: { control: "text" },
+    icon: { control: "text" },
+    includeDate: { control: "boolean" },
+    includeIcon: { control: "boolean" },
+  },
+} satisfies Meta<typeof TimeCard>;
+export type TimeStory = StoryObj<typeof TimeCard>;
+export const TimeExample: TimeStory = {
+  render: Template,
+  args: {},
+};
+
+export const WithoutDateExample: TimeStory = {
+  render: WithoutDate,
+  args: {},
+};