+++ packages/components/src/Shared/Ripples/index.tsx
@@ -8,12 +8,16 @@ import React, {
 } from "react";
 import styled from "@emotion/styled";
 
-export interface RipplesProps {
-  during?: number;
+export interface RipplesProps extends React.ComponentProps<"div"> {
+  /** the animation duration of the ripple @default 600 */
+  duration?: number;
+  /** the color of the ripple, @default rgba(0, 0, 0, .3) */
   color?: string;
+  /** click even to bind to the ripple, @default void */
   onClick?: (ev: React.MouseEvent<HTMLDivElement>) => void;
-  className?: string;
-  children?: React.ReactNode;
+  /** the children of the ripple */
+  children: React.ReactNode;
+  /** the css border radius of the ripple, @default none */
   borderRadius?: CSSProperties["borderRadius"];
 }
 
@@ -44,89 +48,75 @@ const ParentRipple = styled.div<{
     display: inline-flex;
   `}
 `;
-
-export const Ripples = memo((props: Partial<RipplesProps> = {}) => {
-  const [rippleStyle, setRippleStyle] = useState<CSSProperties>({});
-  const timeoutId = useRef<NodeJS.Timeout | null>(null);
-  const options = {
-    duration: 600,
-    color: "rgba(0, 0, 0, .3)",
-    className: "",
-    borderRadius: "none",
-    ...props,
-  };
-
-  const {
-    children,
-    duration,
-    color,
+/** Ripples is a component that can easily add an interactive ripple effect when clicked, simply wrap your component in Ripples and you're good to go! If your component has a border radius, simply pass the same value as a prop to ripples to mask the effect */
+export const Ripples = memo(
+  ({
+    duration = 600,
+    color = "rgba(0, 0, 0, .3)",
+    borderRadius = "none",
     onClick,
-    className,
-    borderRadius,
+    children,
     ...rest
-  } = options;
+  }: RipplesProps) => {
+    const [rippleStyle, setRippleStyle] = useState<CSSProperties>({});
+    const timeoutId = useRef<NodeJS.Timeout | null>(null);
+    useEffect(() => {
+      return () => {
+        // cleanup
+        if (timeoutId.current) clearTimeout(timeoutId.current);
+      };
+    }, []);
 
-  useEffect(() => {
-    return () => {
-      // cleanup
-      if (timeoutId.current) clearTimeout(timeoutId.current);
-    };
-  }, []);
+    const onClickHandler = useCallback(
+      (event: React.MouseEvent<HTMLDivElement>) => {
+        event.stopPropagation();
+        // clear the timeout if exists
+        if (timeoutId.current !== null) clearTimeout(timeoutId.current);
 
-  const onClickHandler = useCallback(
-    (event: React.MouseEvent<HTMLDivElement>) => {
-      event.stopPropagation();
-      // clear the timeout if exists
-      if (timeoutId.current !== null) clearTimeout(timeoutId.current);
+        const { pageX, pageY, currentTarget } = event;
 
-      const { pageX, pageY, currentTarget } = event;
+        const rect = currentTarget.getBoundingClientRect();
 
-      const rect = currentTarget.getBoundingClientRect();
+        const left = pageX - (rect.left + window.scrollX);
+        const top = pageY - (rect.top + window.scrollY);
+        const size = Math.max(rect.width, rect.height);
 
-      const left = pageX - (rect.left + window.scrollX);
-      const top = pageY - (rect.top + window.scrollY);
-      const size = Math.max(rect.width, rect.height);
-
-      setRippleStyle((state) => ({
-        ...state,
-        left,
-        top,
-        opacity: 1,
-        transform: "translate(-50%, -50%)",
-        transition: "initial",
-        backgroundColor: color,
-      }));
-      // start a timeout to scale the ripple
-      timeoutId.current = setTimeout(() => {
         setRippleStyle((state) => ({
           ...state,
-          opacity: 0,
-          transform: `scale(${size / 9})`,
-          transition: `all ${duration}ms`,
+          left,
+          top,
+          opacity: 1,
+          transform: "translate(-50%, -50%)",
+          transition: "initial",
+          backgroundColor: color,
         }));
-        timeoutId.current = null;
-      }, 50);
+        // start a timeout to scale the ripple
+        timeoutId.current = setTimeout(() => {
+          setRippleStyle((state) => ({
+            ...state,
+            opacity: 0,
+            transform: `scale(${size / 9})`,
+            transition: `all ${duration}ms`,
+          }));
+          timeoutId.current = null;
+        }, 50);
 
-      if (typeof onClick === "function") onClick(event);
-    },
-    [color, duration, onClick]
-  );
+        if (typeof onClick === "function") onClick(event);
+      },
+      [color, duration, onClick]
+    );
 
-  return (
-    <ParentRipple borderRadius={borderRadius}>
-      <div
-        {...rest}
-        className={`${className}`.trim()}
-        style={boxStyle}
-        onClick={onClickHandler}
-      >
-        {children}
-        <StyledRipple
-          style={{
-            ...rippleStyle,
-          }}
-        />
-      </div>
-    </ParentRipple>
-  );
-});
+    return (
+      <ParentRipple borderRadius={borderRadius}>
+        <div {...rest} style={boxStyle} onClick={onClickHandler}>
+          {children}
+          <StyledRipple
+            style={{
+              ...rippleStyle,
+            }}
+          />
+        </div>
+      </ParentRipple>
+    );
+  }
+);