+++ packages/components/src/index.ts
@@ -1,5 +1,5 @@
 // ButtonCard
-export { ButtonCard } from "./Cards/ButtonCard";
+export { ButtonCard, StyledButtonCard } from "./Cards/ButtonCard";
 export type { ButtonCardProps } from "./Cards/ButtonCard";
 // SceneCard
 export { SceneCard } from "./Cards/SceneCard";
@@ -13,6 +13,15 @@ export type { RipplesProps } from "./Shared/Ripples";
 // WeatherCard
 export { WeatherCard } from "./Cards/WeatherCard";
 export type { WeatherCardProps } from "./Cards/WeatherCard";
+// Row
+export { Row } from "./Shared/Row";
+export type { RowProps } from "./Shared/Row";
+// column
+export { Column } from "./Shared/Column";
+export type { ColumnProps } from "./Shared/Column";
+// TimeCard
+export { TimeCard } from "./Cards/TimeCard";
+export type { TimeCardProps } from "./Cards/TimeCard";
 
 // ThemeProvider
 export { ThemeProvider } from "../../components/src/ThemeProvider";